# packages/types/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionOutcome": {
        "description": "What the agent reports back after trying an action.",
        "enum": [
          "SUCCESS",
          "FAILED",
          "RETRY",
          "UNVERIFIED",
          "UNSAFE"
        ],
        "title": "ActionOutcome",
        "type": "string"
      },
      "ActionResultRequest": {
        "properties": {
          "artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "outcome": {
            "$ref": "#/components/schemas/ActionOutcome"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "ActionResultRequest",
        "type": "object"
      },
...
      "ActionType": {
        "enum": [
          "SEND_MESSAGE",
          "MARK_ORDER_SHIPPED",
          "PAIR_SESSION",
          "VERIFY_SESSION"
        ],
        "title": "ActionType",
        "type": "string"
      },
...
      "CartsBatch": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/NormalizedCart"
            },
            "title": "Details",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/NormalizedCartSummary"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "CartsBatch",
        "type": "object"
      },
...
      "ClaimRequest": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "lease_seconds": {
            "default": 300,
            "maximum": 3600.0,
            "minimum": 30.0,
            "title": "Lease Seconds",
            "type": "integer"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "ClaimRequest",
        "type": "object"
      },
      "ClaimedAction": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "recovery": {
            "default": false,
            "title": "Recovery",
            "type": "boolean"
          },
          "type": {
            "$ref": "#/components/schemas/ActionType"
          }
        },
        "required": [
          "id",
          "type",
          "payload",
          "attempts",
          "idempotency_key"
        ],
        "title": "ClaimedAction",
        "type": "object"
      },
...
      "ConversationsBatch": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/NormalizedConversation"
            },
            "title": "Details",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/NormalizedConversationSummary"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "ConversationsBatch",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "AUTH_ERROR",
          "NETWORK_ERROR",
          "CARDMARKET_CHANGED",
          "SELECTOR_NOT_FOUND",
          "TIMEOUT",
          "ACTION_FAILED",
          "VERIFICATION_FAILED",
          "DATABASE_ERROR",
          "UNKNOWN"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeartbeatRequest": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "connection_status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "mode": {
            "$ref": "#/components/schemas/AgentMode"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "version",
          "mode",
          "connection_status"
        ],
        "title": "HeartbeatRequest",
        "type": "object"
      },
      "HeartbeatResponse": {
        "properties": {
          "commands": {
            "items": {
              "$ref": "#/components/schemas/AgentCommand"
            },
            "title": "Commands",
            "type": "array"
          },
          "sync_enabled": {
            "title": "Sync Enabled",
            "type": "boolean"
          },
          "sync_interval_seconds": {
            "title": "Sync Interval Seconds",
            "type": "integer"
          }
        },
        "required": [
          "sync_enabled",
          "sync_interval_seconds"
        ],
        "title": "HeartbeatResponse",
        "type": "object"
      },
...
      "OrdersBatch": {
        "properties": {
          "details": {
            "items": {
              "$ref": "#/components/schemas/NormalizedOrder"
            },
            "title": "Details",
            "type": "array"
          },
          "summaries": {
            "items": {
              "$ref": "#/components/schemas/NormalizedOrderSummary"
            },
            "title": "Summaries",
            "type": "array"
          }
        },
        "title": "OrdersBatch",
        "type": "object"
      },
...
      "SessionReport": {
        "properties": {
          "authenticated": {
            "default": false,
            "title": "Authenticated",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "SessionReport",
        "type": "object"
      },
...
      "SyncBatchResult": {
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "records_changed": {
            "title": "Records Changed",
            "type": "integer"
          },
          "records_found": {
            "title": "Records Found",
            "type": "integer"
          }
        },
        "required": [
          "records_found",
          "records_changed",
          "events"
        ],
        "title": "SyncBatchResult",
        "type": "object"
      },
      "SyncCompleteRequest": {
        "properties": {
          "artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "status": {
            "$ref": "#/components/schemas/SyncRunStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "SyncCompleteRequest",
        "type": "object"
      },
...
      "SyncStartRequest": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "entity": {
            "$ref": "#/components/schemas/SyncEntity",
            "default": "FULL"
          },
          "trigger": {
            "default": "schedule",
            "title": "Trigger",
            "type": "string"
          }
        },
        "required": [
          "agent_id"
        ],
        "title": "SyncStartRequest",
        "type": "object"
      },
      "SyncStartResponse": {
        "properties": {
          "granted": {
            "title": "Granted",
            "type": "boolean"
          },
          "known_carts": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Known Carts",
            "type": "object"
          },
          "known_conversations": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Known Conversations",
            "type": "object"
          },
          "known_orders": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Known Orders",
            "type": "object"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "sync_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Id"
          },
          "sync_run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Run Id"
          }
        },
        "required": [
          "granted"
        ],
        "title": "SyncStartResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Cardmarket Companion API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/agent/actions/claim": {
      "post": {
        "operationId": "claim_action_internal_agent_actions_claim_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ClaimedAction"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Claim Action Internal Agent Actions Claim Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Action",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/actions/{action_id}/result": {
      "post": {
        "operationId": "action_result_internal_agent_actions__action_id__result_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "title": "Action Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionResultRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Result",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/heartbeat": {
      "post": {
        "operationId": "heartbeat_internal_agent_heartbeat_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Heartbeat",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/session": {
      "post": {
        "operationId": "report_session_internal_agent_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Report Session",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/sync/start": {
      "post": {
        "operationId": "sync_start_internal_agent_sync_start_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Start",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/sync/{run_id}/carts": {
      "post": {
        "operationId": "sync_carts_internal_agent_sync__run_id__carts_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartsBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Carts",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/sync/{run_id}/complete": {
      "post": {
        "operationId": "sync_complete_internal_agent_sync__run_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Complete",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/sync/{run_id}/conversations": {
      "post": {
        "operationId": "sync_conversations_internal_agent_sync__run_id__conversations_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationsBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Conversations",
        "tags": [
          "internal-agent"
        ]
      }
    },
    "/internal/agent/sync/{run_id}/orders": {
      "post": {
        "operationId": "sync_orders_internal_agent_sync__run_id__orders_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdersBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncBatchResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Orders",
        "tags": [
          "internal-agent"
        ]
      }
    }
  }
}